funcion en data promedio
